Added IDL client program helioidl.pro Major changes for IDL in AsynQueryServiceServlet
git-svn-id: http://svn.code.sf.net/p/helio-vo/code/trunk@1845 5d9e9ddb-692e-4027-9be8-b4f4f3982da7

+++ helio-idlclient/helio-idlclient-provider/src/main/java/eu/heliovo/idlclient/provider/serialize/IdlConverter.java
@@ -107,9 +107,9 @@ public class IdlConverter {
 				else if(pairs.getValue() instanceof int[] )
 				{
 					//System.out.println(pairs.getKey() + " is int array");
-					float[] array = (float[])(pairs.getValue());
+					int[] array = (int[])(pairs.getValue());
 					boolean first = true;
-					for (float item : array)
+					for (int item : array)
 					{
 						if (first) {
 							first = false;
@@ -124,9 +124,9 @@ public class IdlConverter {
 				else if(pairs.getValue() instanceof double[] )
 				{
 					//System.out.println(pairs.getKey() + " is double array");
-					float[] array = (float[])(pairs.getValue());
+					double[] array = (double[])(pairs.getValue());
 					boolean first = true;
-					for (float item : array)
+					for (double item : array)
 					{
 						if (first) {
 							first = false;
